+++ packages/examples/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 All notable changes to this npm module are documented in this file.
 
-## [0.0.0] - 2023-09-05
+## [0.0.0] - 2023-09-06
 
 - Moved langium grammer langugae client and server from [monaco-languageclient](https://github.com/TypeFox/monaco-languageclient) here
+  - Allow to start the editor in both classic and vscode-api mode